Modularized db config script for better separation of concerns. Added db seeding script to package.json.
