+++ server/index.js
@@ -1,7 +1,13 @@
 const express = require('express');
 const app = express();
 const port = 3456;
-var stocks = require('../db/config');
+var db = require('../db/config');
+var stocks = require('../db/stock');
+
+
+// ============================================
+// ESTABLISH API ENDPOINTS
+// ============================================
 
 // GET /stocks returns a list of all stocks
 app.get('/stocks/', (req, res) => {
@@ -11,14 +17,12 @@ app.get('/stocks/', (req, res) => {
 
 // GET /stocks/:ticker/ returns basic info about a stock
 app.get('/stocks/:ticker', (req, res) => {
-  console.log(req.method, req.path, req.body, req.params.ticker);
   stocks.getInfo(req.params.ticker)
     .then(results => { res.send(JSON.stringify(results)); });
 });
 
 // GET /stocks/:ticker/last1dPrices returns relevant price history for a stock
 app.get('/stocks/:ticker/last1dPrices', (req, res) => {
-  console.log(req.method, req.path, req.body, req.params.ticker);
   stocks.get1dPrices(req.params.ticker)
     .then(results => { res.send(JSON.stringify(results)); });
 });
@@ -26,30 +30,31 @@ app.get('/stocks/:ticker/last1dPrices', (req, res) => {
 
 // GET /stocks/:ticker/last1wPrices returns relevant price history for a stock
 app.get('/stocks/:ticker/last1wPrices', (req, res) => {
-  console.log(req.method, req.path, req.body, req.params.ticker);
   stocks.get1wPrices(req.params.ticker)
     .then(results => { res.send(JSON.stringify(results)); });
 });
 
 // GET /stocks/:ticker/last1mPrices returns relevant price history for a stock
 app.get('/stocks/:ticker/last1mPrices', (req, res) => {
-  console.log(req.method, req.path, req.body, req.params.ticker);
   stocks.get1mPrices(req.params.ticker)
     .then(results => { res.send(JSON.stringify(results)); });
 });
 
 // GET /stocks/:ticker/last1yPrices returns relevant price history for a stock
 app.get('/stocks/:ticker/last1yPrices', (req, res) => {
-  console.log(req.method, req.path, req.body, req.params.ticker);
   stocks.get1yPrices(req.params.ticker)
     .then(results => { res.send(JSON.stringify(results)); });
 });
 
 // GET /stocks/:ticker/last5yPrices returns relevant price history for a stock
 app.get('/stocks/:ticker/last5yPrices', (req, res) => {
-  console.log(req.method, req.path, req.body, req.params.ticker);
   stocks.get5yPrices(req.params.ticker)
     .then(results => { res.send(JSON.stringify(results)); });
 });
+
+
+// ============================================
+// START THE SERVER
+// ============================================
 
 app.listen(port, () => console.log('server is listening!'));
